Remove BTC API utils
