--- packages/wallet/detail/holdings/balanceChart/balanceChart.css
@@ -1,50 +1,0 @@
-@import '../../../../../setup/react/app/mixins.css';
-
-.wrapper {
-  display: flex;
-  overflow: initial;
-  padding: 0;
-
-  & .content {
-    display: flex;
-    height: 200px;
-    flex-grow: 1;
-    position: relative;
-    overflow: hidden;
-    border-bottom-left-radius: var(--border-radius-box);
-    border-bottom-right-radius: var(--border-radius-box);
-    padding-left: 0;
-    padding-right: 0;
-  }
-}
-
-.title {
-  @mixin contentLargest semi-bold;
-
-  width: 100%;
-  margin: 0;
-  display: flex;
-  justify-content: flex-start;
-  align-items: flex-start;
-  color: var(--color-maastricht-blue);
-  padding: 0 20px 10px;
-  box-sizing: border-box;
-}
-
-.discreetMode {
-  height: 129px;
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: center;
-  justify-content: center;
-
-  & .icon {
-    width: 30px;
-  }
-
-  & p {
-    @mixin contentLarge;
-
-    color: var(--color-slate-gray);
-  }
-}
